Bug 1687400 - Use uint32_t for Stencil FunctionKey. r=arai CLOSED TREE
The `sourceEnd` doesn't help with uniqueness so don't include in the key in
order to save space. A function may start in beginning of script, so add 1 to
the sourceStart to generate the key.

Differential Revision: https://phabricator.services.mozilla.com/D102246

UltraBlame original commit: 0478698744b1b970c65eefe277903ed457826b06

+++ js/src/vm/Xdr.h
@@ -830,7 +830,7 @@ class XDRIncrementalStencilEncoder : public XDRIncrementalEncoderBase {
   
   
   
-  using FunctionKey = uint64_t;
+  using FunctionKey = uint32_t;
   HashSet<FunctionKey> encodedFunctions_;
 
  public:
